Removed extra checks for english words.
We have an extra check for queries for english words. The check checks that the word is present in static word list that we have.
Due to many simple words not present in word list, this checks fails genuine request too many times.
Now that we have integrated rate limiting in the API, we do not need this extra check and thus avoid false negatives.